Parsed user input, changed insert function to read from double parser(); --> 7) Insert , added quit option for user.

+++ list.cpp
@@ -211,7 +211,7 @@ Out: none
 
 */
 
-void List::insert(int num)
+void List::insert(double num)
 {
 
 /*------ FRONT INSERTION: EMPTY LIST -----*/
